+++ compiler/frontend/bs_builtin_ppx.ml
@@ -87,24 +87,6 @@ let expr_mapper ~async_context ~in_function_def (self : mapper)
   (* Its output should not be rewritten anymore *)
   | Pexp_extension extension ->
     Ast_exp_extension.handle_extension e self extension
-  | Pexp_setinstvar ({txt; loc}, expr) ->
-    if Stack.is_empty Js_config.self_stack then
-      Location.raise_errorf ~loc:e.pexp_loc
-        "This assignment can only happen in object context";
-    let name = Stack.top Js_config.self_stack in
-    if name = "" then
-      Location.raise_errorf ~loc:e.pexp_loc
-        "The current object does not assign a name";
-    let open Ast_helper in
-    self.expr self
-      (Exp.apply ~loc:e.pexp_loc
-         (Exp.ident ~loc {loc; txt = Lident "#="})
-         [
-           ( Nolabel,
-             Exp.send ~loc (Exp.ident ~loc {loc; txt = Lident name}) {loc; txt}
-           );
-           (Nolabel, expr);
-         ])
   | Pexp_constant (Pconst_string (s, Some delim)) ->
     Ast_utf8_string_interp.transform_exp e s delim
   | Pexp_constant (Pconst_integer (s, Some 'l')) ->