+++ compiler/ml/ast_helper.mli
@@ -193,10 +193,6 @@ module Exp : sig
   val constraint_ :
     ?loc:loc -> ?attrs:attrs -> expression -> core_type -> expression
   val send : ?loc:loc -> ?attrs:attrs -> expression -> str -> expression
-  val new_ : ?loc:loc -> ?attrs:attrs -> lid -> expression
-  val setinstvar : ?loc:loc -> ?attrs:attrs -> str -> expression -> expression
-  val override :
-    ?loc:loc -> ?attrs:attrs -> (str * expression) list -> expression
   val letmodule :
     ?loc:loc -> ?attrs:attrs -> str -> module_expr -> expression -> expression
   val letexception :
@@ -207,8 +203,6 @@ module Exp : sig
     expression
   val assert_ : ?loc:loc -> ?attrs:attrs -> expression -> expression
   val lazy_ : ?loc:loc -> ?attrs:attrs -> expression -> expression
-  val poly :
-    ?loc:loc -> ?attrs:attrs -> expression -> core_type option -> expression
   val newtype : ?loc:loc -> ?attrs:attrs -> str -> expression -> expression
   val pack : ?loc:loc -> ?attrs:attrs -> module_expr -> expression
   val open_ :