+++ compiler/ml/ast_iterator.ml
@@ -333,12 +333,6 @@ module E = struct
       sub.expr sub e;
       sub.typ sub t
     | Pexp_send (e, _s) -> sub.expr sub e
-    | Pexp_new lid -> iter_loc sub lid
-    | Pexp_setinstvar (s, e) ->
-      iter_loc sub s;
-      sub.expr sub e
-    | Pexp_override sel ->
-      List.iter (iter_tuple (iter_loc sub) (sub.expr sub)) sel
     | Pexp_letmodule (s, me, e) ->
       iter_loc sub s;
       sub.module_expr sub me;
@@ -348,9 +342,6 @@ module E = struct
       sub.expr sub e
     | Pexp_assert e -> sub.expr sub e
     | Pexp_lazy e -> sub.expr sub e
-    | Pexp_poly (e, t) ->
-      sub.expr sub e;
-      iter_opt (sub.typ sub) t
     | Pexp_newtype (_s, e) -> sub.expr sub e
     | Pexp_pack me -> sub.module_expr sub me
     | Pexp_open (_ovf, lid, e) ->