+++ compiler/ml/pprintast.ml
@@ -706,15 +706,6 @@ and expression ctxt f x =
       in
       let lst = sequence_helper [] x in
       pp f "@[<hv>%a@]" (list (expression (under_semi ctxt)) ~sep:";@;") lst
-    | Pexp_new li -> pp f "@[<hov2>new@ %a@]" longident_loc li
-    | Pexp_setinstvar (s, e) ->
-      pp f "@[<hov2>%s@ <-@ %a@]" s.txt (expression ctxt) e
-    | Pexp_override l ->
-      (* FIXME *)
-      let string_x_expression f (s, e) =
-        pp f "@[<hov2>%s@ =@ %a@]" s.txt (expression ctxt) e
-      in
-      pp f "@[<hov2>{<%a>}@]" (list string_x_expression ~sep:";") l
     | Pexp_letmodule (s, me, e) ->
       pp f "@[<hov2>let@ module@ %s@ =@ %a@ in@ %a@]" s.txt
         (module_expr reset_ctxt) me (expression ctxt) e
@@ -724,12 +715,6 @@ and expression ctxt f x =
         cd (expression ctxt) e
     | Pexp_assert e -> pp f "@[<hov2>assert@ %a@]" (simple_expr ctxt) e
     | Pexp_lazy e -> pp f "@[<hov2>lazy@ %a@]" (simple_expr ctxt) e
-    (* Pexp_poly: impossible but we should print it anyway, rather than
-       assert false *)
-    | Pexp_poly (e, None) -> pp f "@[<hov2>!poly!@ %a@]" (simple_expr ctxt) e
-    | Pexp_poly (e, Some ct) ->
-      pp f "@[<hov2>(!poly!@ %a@ : %a)@]" (simple_expr ctxt) e (core_type ctxt)
-        ct
     | Pexp_open (ovf, lid, e) ->
       pp f "@[<2>let open%s %a in@;%a@]" (override ovf) longident_loc lid
         (expression ctxt) e